pyproject.toml
[tool.poetry]
name = "k8spulse"
version = "v0.6.0"
description = "Kubernetes emergency monitoring tool for generating health reports and insights."
authors = ["Rodrigo Estrada <rodrigo.estrada@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
jinja2 = ">=3.0"
matplotlib = ">=3.4"
pandas = ">=1.3"
openai = ">=0.27.0"
click = ">=8.0"
kubernetes-client = "^0.1.8"
rich = "^13.9.4"
pony = "^0.7.19"

[tool.poetry.scripts]
k8spulse = "k8spulse.cli:cli"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
line-length = 88
target-version = ['py39']
